compare two float32s rather than float32 and float64

+++ geb/hydrology/routing.py
@@ -1402,7 +1402,8 @@ class Routing(Module):
             )
 
             assert (
-                outflow_per_waterbody_m3 <= self.hydrology.waterbodies.var.storage
+                outflow_per_waterbody_m3
+                <= self.hydrology.waterbodies.var.storage.astype(np.float32)
             ).all(), "outflow cannot be smaller or equal to storage"
 
             side_flow_channel_m3_per_hour = (
